🚧 E0Z7PF acr: split validation module
